pyproject.toml
[project]
name = "lla-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "db-dtypes>=1.4.4",
    "google-cloud-bigquery>=3.40.1",
    "ipykernel>=7.2.0",
    "kaleido>=1.2.0",
    "notebook>=7.5.3",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "pytest>=8.4.2",
]
